# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Бэкенд «Умного города» — обращения жителей МКД к управляющей компании, плановые работы, справочники домов/УК/аварийных служб. Один инстанс = одна УК.\n\nРоли: representative (представитель УК), dispatcher (диспетчер), resident (житель). Роль определяется самим пользователем на бэкенде — в запросах отдельно не передаётся.",
        "title": "MAXITO API",
        "contact": {
            "name": "MAXITO"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/max": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход из мини-приложения (initData -\u003e JWT)",
                "operationId": "authLoginByMax",
                "parameters": [
                    {
                        "description": "initData из window.WebApp.initData, как есть",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_common.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_common.loginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "invalid_init_data | init_data_expired",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "not_bound | inactive",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "auth_not_configured",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/internal/bind": {
            "post": {
                "security": [
                    {
                        "InternalKey": []
                    }
                ],
                "description": "Только для сервера бота, требует X-Internal-Key вместо обычного JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Привязать max_user_id к пользователю по телефону",
                "operationId": "botBind",
                "parameters": [
                    {
                        "description": "Телефон + max_user_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_bot.bindRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_bot.bindResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "неверный или отсутствующий X-Internal-Key",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "code: inactive",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "code: phone_not_registered",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "code: conflict — телефон/max_user_id уже привязаны к другому аккаунту",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/internal/issue-token": {
            "post": {
                "security": [
                    {
                        "InternalKey": []
                    }
                ],
                "description": "Только для сервера бота/администратора, требует X-Internal-Key вместо обычного JWT. Не для обычных пользователей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Выпустить JWT по телефону, минуя вход через MAX",
                "operationId": "botIssueToken",
                "parameters": [
                    {
                        "description": "Телефон пользователя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_bot.issueTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_bot.issueTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "неверный или отсутствующий X-Internal-Key",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "code: inactive",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "code: phone_not_registered",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apidoc.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код ошибки; присутствует только у части\nэндпоинтов (см. описание конкретного эндпоинта), у остальных отсутствует.",
                    "type": "string",
                    "example": "bad_request"
                },
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "internal_handlers_bot.bindRequest": {
            "type": "object",
            "required": [
                "max_user_id",
                "phone"
            ],
            "properties": {
                "max_user_id": {
                    "description": "json.Number принимает и число, и строку с числом — бот может прислать\nmax_user_id любым из способов, результат один и тот же.",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "internal_handlers_bot.bindResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string",
                    "example": "resident"
                }
            }
        },
        "internal_handlers_bot.issueTokenRequest": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "type": "string"
                }
            }
        },
        "internal_handlers_bot.issueTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "example": "resident"
                }
            }
        },
...
        "internal_handlers_common.loginRequest": {
            "type": "object",
            "required": [
                "init_data"
            ],
            "properties": {
                "init_data": {
                    "description": "InitData — значение window.WebApp.initData, строка целиком, без изменений.",
                    "type": "string"
                }
            }
        },
        "internal_handlers_common.loginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/internal_handlers_common.loginUser"
                }
            }
        },
        "internal_handlers_common.loginUser": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string",
                    "example": "resident"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT, полученный через POST /auth/max. Передавать как \"Bearer \u0026lt;token\u0026gt;\". Для локальной отладки без MAX бэкенд также принимает X-Max-User-Id (+X-Max-User-Phone при первой привязке), если на сервере включён ALLOW_DEV_HEADERS.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "InternalKey": {
            "description": "Общий секрет между сервером бота и бэкендом — только для /internal/*.",
            "type": "apiKey",
            "name": "X-Internal-Key",
            "in": "header"
        }
    }
}